Faq and add to cart.

+++ add_cart.php
@@ -11,7 +11,7 @@ session_start();
 		<h3>Cart</h3>
 	</div>
 	<?php
-	
+
 	$servername = "localhost";
 	$username = "root";
 	$password = "";
@@ -32,6 +32,7 @@ session_start();
 <body>
 <?php
 	if ($_SERVER["REQUEST_METHOD"] == "POST") {
+		echo "gate 0";
 		$userType = $_SESSION["userType"];
 		$user = $_SESSION["userID"];
 		$sku = $_POST["sku"];
@@ -41,29 +42,59 @@ session_start();
 		$pdo = $conn->query($sql);
 		$result = $pdo->fetch();
 		if (isset($_POST["submit"])) {
+			echo "gate 1";
 			if ($_POST["submit"]=="ADD ITEM") {
+				echo "gate 2";
 				if($result['quantity'] > 0){
-
-					$conn->exec("UPDATE 'inventory'
-					SET 'quantity' = $result[quantity]-1 
+					echo "gate 3";
+					$conn->exec("UPDATE `inventory`
+					SET `quantity` = $result[quantity]-1
 					WHERE sku = $sku;");
+					echo "INSERT INTO `cart` (quantity, sku, userID) VALUES(1, '$sku', '$user')";
 
-					$conn->exec("INSERT INTO cart ('1', 'sku = $sku', 'user = $userID')");
+					$conn->exec("INSERT INTO `cart` (quantity, sku, userID) VALUES(1, '$sku', '$user')");
 
 					$conn->commit();
 				}
 			}
 
 			if ($_POST["submit"]=="REMOVE ITEM") {
-				$conn->exec("UPDATE 'inventory'
-				SET 'quantity' = $result[quantity] + 1
+				$conn->exec("UPDATE `inventory`
+				SET `quantity` = $result[quantity] + 1
 				WHERE sku = $sku;");
 
-				$conn->exec("DELETE FROM cart WHERE sku = $sku;");
+				$conn->exec("DELETE FROM cart WHERE sku = $sku AND userID = $user;");
 				$conn->commit();
 			}
-			
+
 		}
+		echo "<h3>Items in Cart</h3>";
+		$sql = "SELECT `inventory`.`sku`,`inventory`.`category`,`member`.`firstName`,`member`.`lastName`
+						FROM `inventory`,`cart`,`member`
+						WHERE `cart`.`userID` =".$user."
+						AND `inventory`.`sku`=`cart`.`sku`
+						AND `member`.`memberID` = `inventory`.`memberID`";
+		echo "<br>".$sql."<br>";
+		$pdo = $conn->query($sql);
+		echo "<table class=\"w3-striped\">";
+		echo "<tr><th>Style</th><th>Seller</th><th>Name</th><th>        </th><th>      </th><tr>";
+		while($cart = $pdo->fetch()){
+			$sql2 = "SELECT `picURL` FROM `picture` WHERE `sku`=".$cart["sku"]."
+							ORDER BY `picURL` ASC
+							LIMIT 1";
+			$pdo2 = $conn->query($sql2);
+			echo "<tr>";
+			while($pic = $pdo2->fetch()){
+				echo "<td><div class=\"w3-card-8\"><img src=\"".$pic["picURL"]."\" width=\"300\" height=\"300\"></div></td>";
+			}
+			echo "</tr>";
+			echo "<tr><td>".$cart["category"]."</td><td>".$cart["firstName"]."</td><td>".$cart["lastName"]."</td>";
+			echo "<td><form method=\"post\" action=\"add_cart.php\" target=\"_self\">";
+			echo "<input type=\"submit\" value=\"REMOVE ITEM\" name=\"submit\">";
+			echo "<input type=\"number\" name=\"sku\" hidden value=\"".$cart["sku"]."\">";
+			echo "</td></form></tr>";
+		}
+		echo "</table>";
 	}
 	$conn = null;
 ?>
